Move and rename 'ReactInput' to 'ValidatingInput'
